+++ app/Main.hs
@@ -1,10 +1,12 @@
+
+{-# LANGUAGE TypeApplications #-}
+
 module Main where
 
-import qualified Compile as Compile
-import qualified Config as Config
 import Development.GitRev
 import Juvix.Library
 import qualified Juvix.Library.Feedback as Feedback
+import qualified Juvix.Pipeline as Compile
 import Options
 import Options.Applicative
 import System.Directory
@@ -109,17 +111,42 @@ run ctx opt = do
           do (liftIO $ readFile fin)
             >>= Compile.parse
             >>= liftIO . print
-        Typecheck fin backend ->
-          do (liftIO $ readFile fin)
-            >>= Compile.parse
-            >>= Compile.typecheck backend
-            >>= liftIO . print
-        Compile fin fout backend ->
-          do (liftIO $ readFile fin)
-            >>= Compile.parse
-            >>= Compile.typecheck backend
-            >>= Compile.compile backend
-            >>= Compile.writeout fout
-            >>= liftIO . print
+        Typecheck fin (Michelson backend) ->
+          typecheck fin backend
+        Compile fin fout (Michelson backend) ->
+          compile fin fout backend
         Version -> liftIO $ putDoc versionDoc
         _ -> Feedback.fail "Not implemented yet."
+
+compile ::
+  forall b.
+  ( Compile.HasBackend b,
+    Show (Compile.Ty b),
+    Show (Compile.Val b)
+  ) =>
+  FilePath ->
+  FilePath ->
+  b ->
+  Compile.Pipeline ()
+compile fin fout _b = do
+  (liftIO $ readFile fin)
+    >>= Compile.parse
+    >>= Compile.typecheck @b
+    >>= Compile.compile @b
+    >>= Compile.writeout fout
+    >>= liftIO . print
+
+typecheck ::
+  forall b.
+  ( Compile.HasBackend b,
+    Show (Compile.Ty b),
+    Show (Compile.Val b)
+  ) =>
+  FilePath ->
+  b ->
+  Compile.Pipeline ()
+typecheck fin _b = do
+  (liftIO $ readFile fin)
+    >>= Compile.parse
+    >>= Compile.typecheck @b
+    >>= liftIO . print
