+++ src/Juvix/Pipeline/Backend.hs
@@ -1,0 +1,8 @@
+module Juvix.Pipeline.Backend 
+    ( module Juvix.Pipeline.Backend.Internal,
+      module Juvix.Pipeline.Backend.Michelson
+    )
+where
+
+import Juvix.Pipeline.Backend.Internal
+import Juvix.Pipeline.Backend.Michelson
